+++ bindgen-tests/tests/headers/disable-whole-symbol-regex.h
@@ -1,0 +1,6 @@
+// bindgen-flags: --disable-whole-symbol-regex --allowlist-function="^foo"
+
+int foo();
+int foo_bar();
+int bar_foo();
+int bar_foo_baz();